Introduce PushEx.Application for internal dependencies + revamp PushEx Supervision name
