+++ examples/load-test/server/lib/socket_server/application.ex
@@ -7,7 +7,7 @@ defmodule SocketServer.Application do
 
   def start(_type, _args) do
     children = [
-      PushEx.Supervisor,
+      PushEx,
       {SocketServer.StatsLogger,
        [
          collector: SocketServer.Collector.Verbose,
