+++ mix.exs
@@ -25,6 +25,7 @@ defmodule PushEx.MixProject do
   # Type `mix help compile.app` for more information.
   def application do
     [
+      mod: {PushEx.Application, []},
       extra_applications: [:logger, :runtime_tools]
     ]
   end
