+++ test/test_helper.exs
@@ -1,3 +1,3 @@
 PushEx.Test.MockInstrumenter.setup()
-{:ok, _pid} = PushEx.Supervisor.start_link([])
+{:ok, _pid} = Supervisor.start_link([PushEx], strategy: :one_for_one)
 ExUnit.start()
